Major rewrite of redmine.hs
I'd forgotten about this repository!  Oops!  Anyway, it's here
and it's checked in now.  Much of the work done here is in
preparation for subcommands which consume entries in the server
in addition to populating them.

In the process, we've moved from 'wreq' to 'req', a more-modern
web framework (or at least, one with more active development)